+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - allow anonymous function in local scope
     - can access upper scope variables (except for fields, temporary env)
 - support brace expansion in command argument list and source statement
+- add ``xtrace`` runtime option
+    - trace execution of commands
+    - also support ``-x`` command line option
 
 #### Builtin